Update ggml-cuda.cu
Support Linux n1 4.18.0-80.7.1.el8_0.x86_64 #1 SMP Sat Aug 3 15:14:00 UTC 2019 x86_64 x86_64 x86_64 GNU/Linux
whit "NVIDIA-SMI 460.32.03    Driver Version: 460.32.03    CUDA Version: 11.2" & GCC 9.3

Fixed
ggml/src/ggml-cuda.cu(6430): error: more than one instance of overloaded function "log2" matches the argument list:
            function "log2(double)"
            function "log2(float)"
            argument types are: (const int)

1 error detected in the compilation of "/home/u2601933/chatglm.cpp/third_party/ggml/src/ggml-cuda.cu".
make[2]: *** [third_party/ggml/src/CMakeFiles/ggml.dir/ggml-cuda.cu.o] Error 1
make[1]: *** [third_party/ggml/src/CMakeFiles/ggml.dir/all] Error 2
make: *** [all] Error 2

+++ src/ggml-cuda.cu
@@ -7755,7 +7755,7 @@ static void ggml_cuda_op_alibi(
     //GGML_ASSERT(ne01 + n_past == ne00);
     GGML_ASSERT(n_head == ne02);
 
-    const int n_heads_log2_floor = 1 << (int) floor(log2(n_head));
+    const int n_heads_log2_floor = 1 << (int) floor(log2((float)n_head));
 
     const float m0 = powf(2.0f, -(max_bias) / n_heads_log2_floor);
     const float m1 = powf(2.0f, -(max_bias / 2.0f) / n_heads_log2_floor);
